united to_memory_config tests into single file
